+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 * Enhancements
   * Add `known_methods` option to the router
   * Add Mix task to generate a resource module
+  * Add `PhoenixRest.Resource` to wrap `PlugRest.Resource` and provide
+    documentation